Keep experimental hierarchy scoped to the sidebar
