+++ apps/app/src/components/promptbox/NewThreadComposer.tsx
@@ -70,7 +70,6 @@ import {
 } from "@bb/client-core";
 import {
   getProjectComposeRoutePath,
-  getSettingsRoutePath,
   getThreadRoutePath,
   isProjectlessProjectId,
 } from "@/lib/route-paths";
@@ -998,10 +997,6 @@ export function NewThreadComposer({
     },
     [navigate, projectId],
   );
-  const handleNewMachine = useCallback(
-    () => navigate(getSettingsRoutePath("machines")),
-    [navigate],
-  );
   const [commandQuery, setCommandQuery] = useState<string | null>(null);
   const [hasComposerFocused, setHasComposerFocused] = useState(false);
   const handleEditorFocus = useCallback(() => {
@@ -1353,7 +1348,6 @@ export function NewThreadComposer({
               reuseDisabled: reuseThreadOptions.length === 0,
               worktreeDisabledReason,
               disabled: locks.environment,
-              ...(isProjectless ? { onNewMachine: handleNewMachine } : {}),
               ...(!isProjectless && options.onRequestMachineSetup
                 ? { onRequestMachineSetup: options.onRequestMachineSetup }
                 : {}),
@@ -1477,7 +1471,6 @@ export function NewThreadComposer({
       handleCreateBranchFrom,
       handleEditorFocus,
       handleModelChange,
-      handleNewMachine,
       handlePermissionChange,
       handleProjectChange,
       handleProviderChange,
